requirements.txt
langchain-ollama
langchain
langchain_community
python-dotenv
langchain_core
streamlit
